Revert "fix: add install lock to prevent concurrent npm install corruption (#76)"
This reverts commit f9ee2114571980239d7e57540ee86b82fa6515ed.
